doc: disable intersphinx mapping
The intersphinx extensions let one reserve a link namespace and map it
to a remote URL, for example to refer to the Python documentation. A
drawback is that, from time to time, Sphinx will attempt to download an
object inventory from the remote, that is not convenient when running
sphinx in offline mode such as in a plane.

I am assuming most projects will not link to upstream Python
documentation and hence it is safe to remove the intersphinx extension.
The lines are left commented to let one easily reenable the feature
whenever they have a use case of it.

Change-Id: Idef8e88545715a3f1cb9ace002960a7ad3e37b69