Remove .gitreview
This probably needs some explanation.

There are two ways projects get into the OpenStack gerrit
infrastructure.

a) They are declared in openstack-infra/config with an upstream ref.
b) They are declared without an upstream ref.

In (a) .gitreview is incorrect for a cookiecutter cut project - the
project cannot be used with gerrit until it is in the system, and
until the review is approved and actioned the contents of .gitreview
cannot be determined.

In (b) the openstack-infra new project automation creates a correct
.gitreview automatically, which causes pushing the new project
contents to error, as the contents of .gitreview differ (due to no
trailing \n in the cookiecutter output).

So, it seems to me that in both cases we're better served by not
including .gitreview in cookiecutter.

Change-Id: I581cc81fda059d34480579313909f37e448946ea