Fix documentation template issues
This is a roll up of some of the fixes in bash8 that
were needed after doing a cookiecut to create the project.

Update to use oslosphinx instead of oslo.sphinx, fix
the reference to the README.rst file, and add a header
to the contributing doc page, since the default file doesn't
come with one.

Change-Id: Ic3033b7710e9c24ed044d322d08e1eb95e5dc760