Use oslotest rather than copying helpers
The basic test setup is now handled by oslotest and the recommended way
of setting up tests should be to use that.

This theoretically means we should be able to remove testtools from
test-requirements.txt however it has been left as it will likely be
needed later.

Fixtures was removed as a dependency because commonly the required
fixtures would have been set up by oslotest.

Change-Id: I7fa71207ef56ce308fbe4e7a77fa3b9a7e2444ae