Fix Contributing template location
There were several discussions[1][2] on includes and
where the template should actually be living for the U
community goal about contributor docs.

Having the information in CONTRIBUTING.rst so that it
appears on github and using an include in
doc/source/contributor/contributing.rst to pull that text
in via Sphinx is a cleaner way of keeping all the info
together.

[1] https://review.opendev.org/696001
[2] https://review.opendev.org/#/c/707735/

Change-Id: I2bcda1e0f14dac7843869666df4250604b64c265