Minimise info in CONTRIBUTING.rst
This reverts commit ed3a0b07c72667628064dc7c5d8eae47dcae16d6 so that the
CONTRIBUTING.rst file contains only the same information that it
historically had, plus adds links to the canonical OpenDev repo and the
new project-specific contributor documentation.

It also links to the new quick-start page in the Contributor Guide
instead of the OpenStack Infra manual, as the latter is expected to be
moved to OpenDev and scrubbed of OpenStack-specific information.

This has the effect of minimising the amount of information contained in
each repo's CONTRIBUTING file, in favour of keeping it in one place in
the Contributor Guide that can easily be kept up to date.

The content of the new project-specific contributor docs template is
moved back to the docs tree, with some minor cleanups, instead of
transcluding it.

Change-Id: I02dd34a1cf6b62d12f7a26d21b3a2857d82d1b0a
Depends-On: https://review.opendev.org/708939